Change components names
